+++ servo/ports/cef/interfaces/cef_client.rs
@@ -151,7 +151,7 @@ pub struct _cef_client_t {
   
   
   
-  pub ref_count: uint,
+  pub ref_count: usize,
